Move IUserservice.cs to Core.IService
