+++ CS20.FW.WorkSchedule.Core/IService/IUserService.cs
@@ -1,0 +1,14 @@
+using System.Collections.Generic;
+using CS20.FW.WorkSchedule.Core.Model;
+
+namespace CS20.FW.WorkSchedule.Core.IService
+{
+    public interface IUserService
+    {
+        List<User> GetUsers();
+        User GetUserById(int id);
+        User CreateUser(User user);
+        User RemoveUser(int id);
+        User UpdateUser(User user);
+    }
+}
